use querydsl and pagination

+++ src/main/java/com/trong/form/AdvancedSearchForm.java
@@ -5,10 +5,10 @@ import lombok.Data;
 @Data
 public class AdvancedSearchForm {
     private String keyword;
-    private Integer departmentId;
+    private Long departmentId;
     private Boolean requiredFemale;
-    private Integer averageAge;
-    private Integer yearsOfExperience;
+    private Long averageAge;
+    private Long yearsOfExperience;
     private Double salary;
-    private Integer educationalLevelId;
+    private Long educationalLevelId;
 }
